cheat_code/requirements.txt
PyQt5>=5.15
mss>=9.0
pytesseract>=0.3.10
SpeechRecognition>=3.10
PyAudio>=0.2.14
soundcard>=0.4.3
numpy>=1.24,<2.0
pywin32>=306
openai>=1.30.0
keyboard>=0.13.5
Pillow>=10.0
pyinstaller>=6.0
